Split compat.rs into a folder mod with seperate impls for curves
