+++ src/compat/mod.rs
@@ -1,7 +1,8 @@
 use elliptic_curve::{ops::Reduce, point::AffineCoordinates, Curve, CurveArithmetic, PrimeCurve};
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 
-
+pub mod k256_impl;
+pub mod p256_impl;
 
 /// Represents a curve suitable for use in cait-sith.
 ///
@@ -31,78 +32,6 @@ pub trait CSCurve: PrimeCurve + CurveArithmetic {
     ) -> Result<Self::AffinePoint, D::Error>;
 }
 
-#[cfg(any(feature = "k256", test))]
-mod k256_impl {
-    use super::*;
-
-    use elliptic_curve::bigint::Bounded;
-    use k256::Secp256k1;
-
-    impl CSCurve for Secp256k1 {
-        const NAME: &'static [u8] = b"Secp256k1";
-        const BITS: usize = <Self::Uint as Bounded>::BITS;
-
-        fn serialize_point<S: Serializer>(
-            point: &Self::AffinePoint,
-            serializer: S,
-        ) -> Result<S::Ok, S::Error> {
-            point.serialize(serializer)
-        }
-
-        fn deserialize_point<'de, D: Deserializer<'de>>(
-            deserializer: D,
-        ) -> Result<Self::AffinePoint, D::Error> {
-            Self::AffinePoint::deserialize(deserializer)
-        }
-    }
-}
-
-#[cfg(any(feature = "p256", test))]
-mod p256_impl {
-    use super::*;
-
-    use elliptic_curve::bigint::Bounded;
-    use p256::NistP256;
-
-    impl CSCurve for NistP256 {
-        const NAME: &'static [u8] = b"Secp256r1";
-        const BITS: usize = <Self::Uint as Bounded>::BITS;
-
-        fn serialize_point<S: Serializer>(
-            point: &Self::AffinePoint,
-            serializer: S,
-        ) -> Result<S::Ok, S::Error> {
-            point.serialize(serializer)
-        }
-
-        fn deserialize_point<'de, D: Deserializer<'de>>(
-            deserializer: D,
-        ) -> Result<Self::AffinePoint, D::Error> {
-            Self::AffinePoint::deserialize(deserializer)
-        }
-    }
-}
-
-
-#[cfg(test)]
-mod test_scalar_hash {
-    use super::*;
-
-    use digest::{Digest, FixedOutput};
-    use ecdsa::hazmat::DigestPrimitive;
-    use elliptic_curve::{ops::Reduce, Curve};
-    use k256::{FieldBytes, Scalar, Secp256k1};
-
-    #[cfg(test)]
-    pub(crate) fn scalar_hash(msg: &[u8]) -> <Secp256k1 as CurveArithmetic>::Scalar {
-        let digest = <Secp256k1 as DigestPrimitive>::Digest::new_with_prefix(msg);
-        let m_bytes: FieldBytes = digest.finalize_fixed();
-        <Scalar as Reduce<<Secp256k1 as Curve>::Uint>>::reduce_bytes(&m_bytes)
-    }
-}
-
-#[cfg(test)]
-pub(crate) use test_scalar_hash::scalar_hash;
 
 #[derive(Clone, Copy)]
 pub(crate) struct SerializablePoint<C: CSCurve>(C::AffinePoint);
@@ -140,3 +69,4 @@ impl<'de, C: CSCurve> Deserialize<'de> for SerializablePoint<C> {
 pub(crate) fn x_coordinate<C: CSCurve>(point: &C::AffinePoint) -> C::Scalar {
     <C::Scalar as Reduce<<C as Curve>::Uint>>::reduce_bytes(&point.x())
 }
+
